Initial commit
I started writing this after hearing about the "This Face Does Not Exist" GAN project.

I was going to scrape listings to build a corpus of text so that I could then generate artificial descriptions of houses.

But this is basically three-part project:
1. Get the list of local listings
2. Iterate over each individual listing to scrape the text
3. Simulate text

This script is the first part of that project.

It's actually adapted from another one of my scripts that scrapes Amazon reviews by ASIN.